+++ src/commands/mod.rs
@@ -1,6 +1,7 @@
 mod ci;
 mod fix;
 mod fix_unsafe;
+mod full;
 pub mod init_config;
 mod lint;
 mod pitstop;
@@ -16,6 +17,7 @@ use crate::stacks;
 pub use ci::ci;
 pub use fix::fix;
 pub use fix_unsafe::fix_unsafe;
+pub use full::full;
 pub use init_config::init_config;
 pub use lint::lint;
 pub use pitstop::pitstop;