+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - `ci --test` runs a subset of the configured tests, e.g. `trident ci --test=unit+cuke`
 - `pitstop --test` runs a subset of the configured tests, e.g. `trident pitstop --test=unit+cuke`
 - `test --test` runs a subset of the configured tests, e.g. `trident test --test=unit+cuke`
+- new `full` command runs all lints, fixes, and tests on all files
+- `full --test` runs a subset of the configured tests, e.g. `trident full --test=unit+cuke`
 - renamed to Trident
 
 ## 0.1.0